+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ProjectPackagePrefixes.java
@@ -35,7 +35,6 @@ import java.util.regex.Pattern;
 import edu.umd.cs.findbugs.ba.AnalysisContext;
 import edu.umd.cs.findbugs.charsets.UTF8;
 import edu.umd.cs.findbugs.internalAnnotations.DottedClassName;
-import edu.umd.cs.findbugs.util.Util;
 
 /**
  * @author pwilliam
@@ -208,9 +207,7 @@ public class ProjectPackagePrefixes {
         URL u = DetectorFactoryCollection.getCoreResource("projectPaths.properties");
         if (u != null) {
 
-            BufferedReader in = null;
-            try {
-                in = UTF8.bufferedReader(u.openStream());
+            try (BufferedReader in = UTF8.bufferedReader(u.openStream())) {
                 while (true) {
                     String s = in.readLine();
                     if (s == null) {
@@ -224,8 +221,6 @@ public class ProjectPackagePrefixes {
             } catch (IOException e1) {
 
                 AnalysisContext.logError("Error loading projects paths", e1);
-            } finally {
-                Util.closeSilently(in);
             }
 
         }