+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/SourceFinder.java
@@ -43,6 +43,7 @@ import java.util.zip.ZipFile;
 import java.util.zip.ZipInputStream;
 
 import javax.annotation.WillClose;
+import javax.annotation.WillCloseWhenClosed;
 
 import edu.umd.cs.findbugs.Project;
 import edu.umd.cs.findbugs.SourceLineAnnotation;
@@ -55,7 +56,7 @@ import edu.umd.cs.findbugs.util.Util;
  * which is like a classpath, but for finding source files instead of class
  * files.
  */
-public class SourceFinder {
+public class SourceFinder implements AutoCloseable {
     private static final boolean DEBUG = SystemProperties.getBoolean("srcfinder.debug");
 
     private static final int CACHE_SIZE = 50;
@@ -85,12 +86,15 @@ public class SourceFinder {
     /**
      * A repository of source files.
      */
-    private interface SourceRepository {
+    private interface SourceRepository extends AutoCloseable {
         public boolean contains(String fileName);
 
         public boolean isPlatformDependent();
 
         public SourceFileDataSource getDataSource(String fileName);
+
+        @Override
+        void close() throws IOException;
     }
 
     /**
@@ -130,6 +134,11 @@ public class SourceFinder {
 
         private String getFullFileName(String fileName) {
             return baseDir + File.separator + fileName;
+        }
+
+        @Override
+        public void close() {
+            // this class keeps no file handler
         }
     }
 
@@ -208,6 +217,11 @@ public class SourceFinder {
         public boolean isPlatformDependent() {
             return false;
         }
+
+        @Override
+        public void close() {
+            // this class keeps no file handler
+        }
     }
 
     SourceRepository makeInMemorySourceRepository(final String url) {
@@ -245,28 +259,33 @@ public class SourceFinder {
         file.deleteOnExit();
         final BlockingSourceRepository r = new BlockingSourceRepository();
         Util.runInDameonThread(() -> {
-            InputStream in = null;
-            OutputStream out = null;
-            try {
-                URLConnection connection = new URL(url).openConnection();
-                if (getProject().isGuiAvaliable()) {
-                    int size = connection.getContentLength();
-                    in = getProject().getGuiCallback().getProgressMonitorInputStream(connection.getInputStream(), size,
-                            "Loading source via url");
-                } else {
-                    in = connection.getInputStream();
-                }
-                out = new FileOutputStream(file);
+            try (InputStream in = open(url); OutputStream out = new FileOutputStream(file);) {
                 IO.copy(in, out);
                 r.setBase(new ZipSourceRepository(new ZipFile(file)));
             } catch (IOException e) {
                 assert true;
-            } finally {
-                Util.closeSilently(in);
-                Util.closeSilently(out);
             }
         }, "Source loading thread");
         return r;
+    }
+
+    /**
+     * @param url
+     * @return
+     * @throws IOException
+     * @throws MalformedURLException
+     */
+    private InputStream open(final String url) throws IOException, MalformedURLException {
+        InputStream in = null;
+        URLConnection connection = new URL(url).openConnection();
+        if (getProject().isGuiAvaliable()) {
+            int size = connection.getContentLength();
+            in = getProject().getGuiCallback().getProgressMonitorInputStream(connection.getInputStream(), size,
+                    "Loading source via url");
+        } else {
+            in = connection.getInputStream();
+        }
+        return in;
     }
 
     static class BlockingSourceRepository implements SourceRepository {
@@ -281,7 +300,7 @@ public class SourceFinder {
             return ready.getCount() == 0;
         }
 
-        public void setBase(SourceRepository base) {
+        public void setBase(@WillCloseWhenClosed SourceRepository base) {
             this.base = base;
             ready.countDown();
         }
@@ -312,6 +331,10 @@ public class SourceFinder {
             return base.isPlatformDependent();
         }
 
+        @Override
+        public void close() throws IOException {
+            base.close();
+        }
     }
 
     /**
@@ -320,7 +343,7 @@ public class SourceFinder {
     static class ZipSourceRepository implements SourceRepository {
         ZipFile zipFile;
 
-        public ZipSourceRepository(ZipFile zipFile) {
+        public ZipSourceRepository(@WillCloseWhenClosed ZipFile zipFile) {
             this.zipFile = zipFile;
         }
 
@@ -337,6 +360,11 @@ public class SourceFinder {
         @Override
         public SourceFileDataSource getDataSource(String fileName) {
             return new ZipSourceFileDataSource(zipFile, fileName);
+        }
+
+        @Override
+        public void close() throws IOException {
+            zipFile.close();
         }
     }
 
@@ -578,5 +606,11 @@ public class SourceFinder {
         cache = new Cache();
         setSourceBaseList(project.getResolvedSourcePaths());
     }
-}
 
+    @Override
+    public void close() {
+        for (SourceRepository repo : repositoryList) {
+            IO.close(repo);
+        }
+    }
+}