+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/bcp/FieldAccess.java
@@ -36,7 +36,7 @@ import edu.umd.cs.findbugs.ba.vna.ValueNumberFrame;
  * @see Load
  * @see Store
  */
-public abstract class FieldAccess extends SingleInstruction implements org.apache.bcel.Constants {
+public abstract class FieldAccess extends SingleInstruction {
     private final String fieldVarName;
 
     private final String valueVarName;