+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/npe/IsNullValueFrame.java
@@ -250,8 +250,14 @@ public class IsNullValueFrame extends Frame<IsNullValue> {
             return false;
         }
         IsNullValueFrame o2 = (IsNullValueFrame) other;
-        return Objects.equals(decision, o2.decision)
-            && !(trackValueNumbers && !Objects.equals(knownValueMap, o2.knownValueMap));
+        if (!Objects.equals(decision, o2.decision)) {
+            return false;
+        }
+        if (trackValueNumbers && !Objects.equals(knownValueMap, o2.knownValueMap)) {
+            return false;
+        }
+
+        return true;
     }
 
     @Override
@@ -300,4 +306,3 @@ public class IsNullValueFrame extends Frame<IsNullValue> {
         }
     }
 }
-