+++ CHANGELOG.md
@@ -5,8 +5,65 @@ This is the changelog for SpotBugs. This follows [Keep a Changelog v1.0.0](http:
 Currently the versioning policy of this project follows [Semantic Versioning v2.0.0](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - 2018-??-??
-
 * Start migrating STDOUT/STDERR usage to a logging framework
+
+## 3.1.9 - 2018-11-20
+
+### Fixed
+* Fix some out-of-bounds reports from LGTM
+* Update asm to 7.0 for better Java 11 support ([#785](https://github.com/spotbugs/spotbugs/pull/785))
+* Ignore @FXML annotated fields in UR\_UNIT\_READ ([#702](https://github.com/spotbugs/spotbugs/issues/702))
+
+### CHANGED
+* Allow parallel workspace builds in Eclipse with Spotbugs installed
+
+## 3.1.8 - 2018-10-16
+
+### Fixed
+* Update asm to 6.2.1 for better Java 12 support ([#741](https://github.com/spotbugs/spotbugs/issues/741))
+* Fix hash code collision ([#751](https://github.com/spotbugs/spotbugs/pull/751))
+* Partially revert [#688](https://github.com/spotbugs/spotbugs/pull/688) because of the error in specific case with `checkcast` opcode ([#760](https://github.com/spotbugs/spotbugs/pull/760))
+
+## 3.1.7 - 2018-09-12
+
+### Fixed
+* Don't print exit code related output if '-quiet' is passed ([#714](https://github.com/spotbugs/spotbugs/pull/714))
+* Don't underflow the stack at INVOKEDYNAMIC when modeling stack frame types ([#500](https://github.com/spotbugs/spotbugs/issues/500))
+
+### CHANGED
+* ASM_VERSION=ASM7_EXPERIMENTAL by default to support Java 11
+* Removed dependency to jFormatString (GPL) code ([#725](https://github.com/spotbugs/spotbugs/issues/725))
+* Read User Preferences exported from SpotBugs Eclipse Plugin  ([#728](https://github.com/spotbugs/spotbugs/issues/728))
+
+### ADDED
+* Set ASM_VERSION=ASM6 if system property spotbugs.experimental=false
+
+## 3.1.6 - 2018-07-18
+
+### Fixed
+
+* Potential NPE in test-harness-core ([#671](https://github.com/spotbugs/spotbugs/issues/671))
+* Support project path with spaces in test-harness-core ([#683](https://github.com/spotbugs/spotbugs/issues/683))
+* Processing of "J" (long value constants) was not processed in `OpcodeStack.Item(OpcodeStack.Item, String)`
+* Processing of "Z" (boolean value constants) was not processed in `OpcodeStack.Item(OpcodeStack.Item, String)`
+* Processing of Box classes like `java.lang.Integer` was not processed in `OpcodeStack.Item(OpcodeStack.Item, String)`
+
+## 3.1.5 - 2018-06-15
+
+### Fixed
+
+* Keep IO.close(Closeable) that was deleted by 3.1.4 ([#661](https://github.com/spotbugs/spotbugs/issues/661))
+
+## 3.1.4 - 2018-06-11 [YANKED]
+
+### Fixed
+
+* RANGE_ARRAY_LENGTH and RANGE_ARRAY_OFFSET false negative ([#595](https://github.com/spotbugs/spotbugs/issues/595))
+* Close source file after analysis ([#591](https://github.com/spotbugs/spotbugs/issues/591))
+* Inconsistent reporting for EI_EXPOSE_REP2 ([#603](https://github.com/spotbugs/spotbugs/issues/603))
+* Update asm to 6.2 for better Java 11 support ([#648](https://github.com/spotbugs/spotbugs/issues/648))
+* False positive: 'return value ignored' on Guavas Preconditions.checkNotNull() ([#578](https://github.com/spotbugs/spotbugs/issues/578))
+* spotbugs-ant Ant dependency in wrong scope ([#655](https://github.com/spotbugs/spotbugs/issues/655))
 
 ## 3.1.3 - 2018-04-18